fixed code

+++ app.py
@@ -11,14 +11,16 @@ import json
 import os
 import google.oauth2.credentials
 import googleapiclient.discovery
+from flask_migrate import Migrate
 
 
 app = Flask(__name__)
-app.config['GOOGLE_ID'] = "17614516131-qf2o1gn2cop4l200agt15ke45r9hqqgf.apps.googleusercontent.com"  # noqa
+app.config['GOOGLE_ID'] = "17614516131-qf2o1gn2cop4l200agt15ke45r9hqqgf.apps.googleusercontent.com"
 app.config['GOOGLE_SECRET'] = "-bTgr4zCGgDEk7gaGd1yZmBF"
 app.debug = True
 app.secret_key = 'development'
 oauth = OAuth(app)
+
 
 google = oauth.remote_app(
     'google',
@@ -34,17 +36,15 @@ google = oauth.remote_app(
     authorize_url='https://accounts.google.com/o/oauth2/auth',
 )
 
-
 def login_required(f):
     @wraps(f)
     def wrap(*args, **kwargs):
-        if 'logged_in' in session and session['logged_in'] is True:
+        if 'logged_in' in session and session['logged_in'] == True:
             return f(*args, **kwargs)
-        else:
+        else:            
             return redirect(url_for('login'))
 
     return wrap
-
 
 # ======== Routing ========== #
 # -------- Login ------------ #
@@ -64,8 +64,10 @@ def login():
             return json.dumps({'status': 'Both fields required'})
         return render_template('login.html', form=form)
     user = helpers.get_user()
+    session['logged_in'] = True
+    session['username'] = user.username
+    session['userid'] = user.id     
     return redirect(url_for('showcatalogs'))
-
 
 @app.route('/loginGoogle')
 def loginGoogle():
@@ -82,8 +84,19 @@ def authorized():
         )
     session['google_token'] = (resp['access_token'], '')
     me = google.get('userinfo')
+    data = me.data 
+    email = data['email']
     session['logged_in'] = True
-    return redirect(url_for('showcatalogs'))
+    session['username'] = email
+    if helpers.is_registered(email) == 0:
+        password = helpers.hash_password(email)
+        helpers.add_user(email, password, email)            
+    user = helpers.get_user_by_email(email)
+    id = user.id 
+    session['userid'] = id 
+    return redirect(url_for('showcatalogs'))    
+
+
 
 
 @google.tokengetter
@@ -94,6 +107,8 @@ def get_google_oauth_token():
 @app.route('/showcatalogs', methods=['GET', 'POST'])
 @login_required
 def showcatalogs():
+    #if the request is not ajax
+    # then render page
     if not request.is_xhr:
         catalog = helpers.query_catalog()
         item_details = helpers.query_catalog_and_item()
@@ -113,6 +128,18 @@ def showcatalogs():
 @login_required
 def get_tasks():
     item_details = helpers.query_catalog_and_item()
+    ret = []
+    for item in item_details:
+        ret.append({"item": item.item_name, "category": item.catalog_name})
+    return jsonify({'Category': ret})
+
+
+@app.route('/queryitem/api/v1.0/item/<item_id>', methods=['GET'])
+@login_required
+def get_item(item_id):
+    item_details = helpers.query_item(item_id)
+    if (len(item_details) == 0):
+        return jsonify({'Category': "empty"})
     ret = []
     for item in item_details:
         ret.append({"item": item.item_name, "category": item.catalog_name})
@@ -140,35 +167,47 @@ def additem():
         item_text = request.form['item_text'].lower()
         description_text = request.form['description_text'].lower()
         id = request.form['categoryId']
-        helpers.add_item(item_text, description_text, id)
+        user_id = session['userid']        
+        helpers.add_item(item_text, description_text, id, user_id)
         return redirect(url_for('showcatalogs'))
 
 
 @app.route('/showitem/<item_id>', methods=['GET', 'POST'])
 @login_required
 def showedititem(item_id):
-    if request.method == 'GET':
-        item_id = item_id
-        item = helpers.query_item(item_id)
-        return render_template('showitem.html', item=item[0])
-    elif request.method == 'POST' and 'action' in request.form and request.form['action'] == 'save':  # noqa
-        item_id = item_id
+    if request.method == 'GET':        
+        item = helpers.query_item(item_id)        
+        return render_template('showitem.html', item=item[0])        
+    elif request.method == 'POST' and 'action' in request.form and request.form['action'] == 'save':  # noqa        
+        # save action
         item_text_value = request.form['item_text']
         description_text = request.form['description_text']
-        helpers.update_item(item_id, item_text_value, description_text)
-        return redirect(url_for('showcatalogs'))
+        user_id = session['userid']             
+        item = helpers.query_item(item_id)[0]
+        if (item.user_id == user_id) :
+           helpers.update_item(item_id, item_text_value, description_text)
+           return redirect(url_for('showcatalogs'))
+        else:
+            # no permission to do update
+           return render_template('message.html', msg = "update")      
     elif request.method == 'POST' and 'action' in request.form and request.form['action'] == 'delete':  # noqa
-        item_id = item_id
-        helpers.delete_item(item_id)
-        return redirect(url_for('showcatalogs'))
+        # delete action
+        user_id = session['userid']             
+        item = helpers.query_item(item_id)[0]     
+        if (item.user_id == user_id) :
+            helpers.delete_item(item_id)
+            return redirect(url_for('showcatalogs'))
+        else:
+            return render_template('message.html' , msg = "delete")                  
     else:
         return redirect(url_for('showcatalogs'))
 
 
 @app.route("/logout")
 def logout():
-    session['logged_in'] = False
-    session.pop('google_token', None)
+    session['logged_in'] = False    
+    session.pop('google_token', None)       
+    session.clear()
     return redirect(url_for('login'))
 
 
@@ -185,7 +224,7 @@ def signup():
                 if not helpers.username_taken(username):
                     helpers.add_user(username, password, email)
                     session['logged_in'] = True
-                    session['username'] = username
+                    session['username'] = username                    
                     return json.dumps({'status': 'Signup successful'})
                 return json.dumps({'status': 'Username taken'})
             return json.dumps({'status': 'User/Pass required'})
@@ -212,4 +251,5 @@ def settings():
 # ======== Main ============ #
 if __name__ == "__main__":
     app.secret_key = os.urandom(12)  # Generic key for dev purposes only
-    app.run(host='localhost', port=8000, debug=True, use_reloader=True)
+    #app.run(debug=True, use_reloader=True)
+    app.run(host='localhost', port=8000,debug=True, use_reloader=True)
